Add IAction interface
This interface will be used for every action a player can take in the
game. The Display method displays the result of the action in the
BoardView, while the Undo method reverses what the Display method does.
The Apply method finalizes the Action and puts the changes into the
Board class.